# docs/api/openapi-generated.yaml
components:
  schemas:
...
    ApiKeyCreatedResponse:
      description: API key created response (includes raw key, shown only once).
      properties:
        apiKey:
          type: string
        createdAt:
          type: string
        id:
          format: uuid
          type: string
        name:
          type: string
        permissions: {}
      required:
      - id
      - name
      - apiKey
      - permissions
      - createdAt
      type: object
    ApiKeyResponse:
      description: API key response item (without raw key).
      properties:
        createdAt:
          type: string
        createdBy:
          format: uuid
          type: string
        id:
          format: uuid
          type: string
        lastUsedAt:
          type:
          - string
          - 'null'
        name:
          type: string
        permissions: {}
        projectId:
          format: uuid
          type: string
      required:
      - id
      - projectId
      - name
      - permissions
      - createdBy
      - createdAt
      type: object
...
    AuditLogResponse:
      description: Audit log response item.
      properties:
        action:
          type: string
        actorId:
          format: uuid
          type:
          - string
          - 'null'
        actorType:
          type: string
        contextId:
          format: uuid
          type:
          - string
          - 'null'
        contextType:
          type:
          - string
          - 'null'
        createdAt:
          type: string
        details: {}
        id:
          format: uuid
          type: string
        resourceId:
          format: uuid
          type: string
        resourceType:
          type: string
      required:
      - id
      - actorType
      - action
      - resourceType
      - resourceId
      - details
      - createdAt
      type: object
...
    CreateApiKeyRequest:
      description: Create API key request body.
      properties:
        name:
          type: string
        permissions: {}
      required:
      - name
      - permissions
      type: object
...
    ListApiKeysResponse:
      description: List API keys response.
      properties:
        data:
          items:
            $ref: '#/components/schemas/ApiKeyResponse'
          type: array
      required:
      - data
      type: object
    ListAuditLogsResponse:
      description: List audit logs response.
      properties:
        data:
          items:
            $ref: '#/components/schemas/AuditLogResponse'
          type: array
        nextCursor:
          type:
          - string
          - 'null'
      required:
      - data
      type: object
...
    ProblemDetails:
      allOf:
      - description: RFC 7807 extension fields (e.g. for 409 `StaleConversation` extra data).
        type:
        - object
        - 'null'
      - properties:
          detail:
            description: Detailed error description.
            type:
            - string
            - 'null'
          errors:
            description: Field-level validation errors.
            items:
              $ref: '#/components/schemas/ValidationError'
            type:
            - array
            - 'null'
          instance:
            description: URI of the specific resource where the error occurred.
            type:
            - string
            - 'null'
          status:
            description: HTTP status code.
            format: int32
            minimum: 0
            type: integer
          title:
            description: Human-readable error summary.
            type: string
          type:
            description: Error type URI for programmatic identification.
            type: string
        required:
        - type
        - title
        - status
        type: object
      description: RFC 7807 Problem Details error response.
...
    ValidationError:
      description: A single field validation error.
      properties:
        code:
          description: Error code for programmatic handling (e.g. "required", "`too_long`").
          type: string
        field:
          description: Field path (e.g. "name", "config.apiKey").
          type: string
        message:
          description: Human-readable error message.
          type: string
      required:
      - field
      - message
      - code
      type: object
...
  securitySchemes:
    bearer_auth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  description: AI-assisted conference/event project management system
  license:
    name: ''
  title: Conf-Ops API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/organizations/{orgId}/audit-logs:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on database failure.
      operationId: list_org_audit_logs
      parameters:
      - description: Organization ID
        in: path
        name: orgId
        required: true
        schema:
          format: uuid
          type: string
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          format: uuid
          type: string
      - description: Max items to return
        in: query
        name: limit
        required: false
        schema:
          format: int64
          type: integer
      - description: Filter by actor type
        in: query
        name: actorType
        required: false
        schema:
          type: string
      - description: Filter by action
        in: query
        name: action
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuditLogsResponse'
          description: Organization audit logs
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List audit logs for an organization.
      tags:
      - audit
...
  /api/v1/projects/{projectId}/api-keys:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on database failure.
      operationId: list_api_keys
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListApiKeysResponse'
          description: API key list
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List API keys for a project.
      tags:
      - api-keys
    post:
      description: |-
        # Errors

        Returns `ProblemDetails` on validation or database failure.
      operationId: create_api_key
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKeyRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreatedResponse'
          description: API key created (raw key shown only once)
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Create an API key for a project.
      tags:
      - api-keys
  /api/v1/projects/{projectId}/api-keys/{keyId}:
    delete:
      description: |-
        # Errors

        Returns `ProblemDetails` on not found.
      operationId: delete_api_key
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: API Key ID
        in: path
        name: keyId
        required: true
        schema:
          format: uuid
          type: string
      responses:
        '204':
          description: API key revoked
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: Revoke (delete) an API key.
      tags:
      - api-keys
  /api/v1/projects/{projectId}/audit-logs:
    get:
      description: |-
        # Errors

        Returns `ProblemDetails` on database failure.
      operationId: list_project_audit_logs
      parameters:
      - description: Project ID
        in: path
        name: projectId
        required: true
        schema:
          format: uuid
          type: string
      - description: Cursor for pagination
        in: query
        name: cursor
        required: false
        schema:
          format: uuid
          type: string
      - description: Max items to return
        in: query
        name: limit
        required: false
        schema:
          format: int64
          type: integer
      - description: Filter by actor type
        in: query
        name: actorType
        required: false
        schema:
          type: string
      - description: Filter by action
        in: query
        name: action
        required: false
        schema:
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuditLogsResponse'
          description: Project audit logs
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
          description: ''
      summary: List audit logs for a project.
      tags:
      - audit
...